// SulambiVosaMain/Technology Transfer _ Sulambi VMS/Source Code/sulambi-frontend-main/sulambi-frontend-main/src/interface/types.ts
export interface EvaluationDataType {
  criteria: any;
  comment: string;
  q13: string;
  q14: string;
}

export interface RequirementsDataType {
  id: number;
  eventId: any;
  medCert: string;
  waiver: string;

  fullname: string;
  email: string;
  srcode: string;
  age: string;
  birthday: string;
  sex: string;
  campus: string;
  collegeDept: string;
  yrlevelprogram: string;
  address: string;
  contactNum: string;
  fblink: string;

  curriculum?: string;
  destination?: string;
  firstAid?: string;
  fees?: string;
  personnelInCharge?: string;
  personnelRole?: string;
  accepted?: number;
}
// ...
export interface EventEvalListType {
  requirements: RequirementsDataType;
  evaluation: EvaluationDataType;
}